requirements.txt
# Core ML/AI Dependencies
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
datasets>=2.14.0

# Tamil Text Processing
unicodedata2>=15.1.0
indic-transliteration>=2.3.0
# polyglot>=16.7.4  # Commented out due to Windows compatibility issues

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
httpx>=0.25.0
aiofiles>=23.2.0

# Logging and Monitoring
loguru>=0.7.2
rich>=13.6.0

# Database Support
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymongo>=4.5.0         # MongoDB
redis>=4.6.0           # Redis for caching

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0

# Optional: For advanced features
sentence-transformers>=2.2.0  # For RAG implementation
openai-whisper>=20230918      # For speech-to-text
TTS>=0.20.0                   # For text-to-speech
